Issue #618 Clean up
