+++ src/sagas/analytics/watch_analytics.ts
@@ -1,12 +1,12 @@
-// tslint:disable:no-expression-statement
-
-import { ForkEffect, takeLatest } from 'redux-saga/effects';
-import { sendAnalyticsData } from './send_analytics_data';
+// tslint:disable:no-expression-statement no-any
 
 import * as constants from '../../application/constants';
+import { AnalyticsAsync } from '../../stores/analytics';
 import { ChooseAnswerAction } from '../../stores/questionnaire';
 import { RouteChangedAction } from '../../stores/router_actions';
 import { AddToSavedListAction } from '../../stores/topics';
+import { ForkEffect, takeLatest, call, CallEffect, PutEffect, put } from 'redux-saga/effects';
+import * as events from './events';
 
 export type WatchedAction = RouteChangedAction | ChooseAnswerAction | AddToSavedListAction;
 
@@ -18,4 +18,28 @@ export function* watchAnalytics(): IterableIterator<ForkEffect> {
             constants.ADD_TO_SAVED_TOPICS,
         ],
         sendAnalyticsData);
+}
+
+type AnalyticsActions = IterableIterator<CallEffect | PutEffect<AnalyticsAsync.SuccessAction | AnalyticsAsync.FailureAction>>;
+
+function* sendAnalyticsData(action: WatchedAction): AnalyticsActions {
+    try {
+        yield call(sendAnalyticsDataAsync, action);
+        yield put(AnalyticsAsync.success());
+    } catch (error) {
+        console.error(`Failed to send analytics data (${error.message})`);
+        yield put(AnalyticsAsync.failure(error.message));
+    }
+}
+
+async function sendAnalyticsDataAsync(action: WatchedAction): Promise<void> {
+    if (action.type === constants.ROUTE_CHANGED) {
+        events.sendScreenHit(action);
+    }
+    if (action.type === constants.CHOOSE_ANSWER) {
+        events.sendAnswerChosenEvent(action.payload.answerId);
+    }
+    if (action.type === constants.ADD_TO_SAVED_TOPICS) {
+        events.sendBookmarkAddedEvent(action.payload.topicId);
+    }
 }
